+++ HISTORY.md
@@ -1,3 +1,21 @@
+# v0.8.0 / 2021-06-23
+
+* Adjustments to support Elasticsearch v7 and up (<https://www.elastic.co/guide/en/elasticsearch/reference/current/removal-of-types.html>)
+  * Updated indices.putMapping: removed `/_mapping/_type` from resource and eliminated `_type` parameter requirement
+  * Updated core.add (core.index): added `/_doc` to path and eliminated `_type` parameter requirement
+  * Updated core.delete: added `/_doc` to path
+  * Updated core.get: added `/_doc` to path and eliminated `_type` parameter requirement
+  * Removed core.moreLikeThis: is now accomplished via search query (<https://www.elastic.co/guide/en/elasticsearch/reference/7.13/query-dsl-mlt-query.html>)
+  * Updated core.multiGet: removed `_type` from path and eliminated `_type` parameter requirement
+  * Updated core.update: removed `_type` from path and eliminated `_type` parameter requirement
+  * Updated core.bulkIndex: removed `_type` from payload interpolation
+  * Updated core.explain: removed `_type` from path and eliminated `_type` parameter requirement
+  * Updated core.search: removed `_type` from path
+  * Updated core.suggest: removed `_type` from path
+  * Updated core.count: removed `_type` from path
+  * Updated core.multiSearch: removed `_type` from path
+  * Updated core.deleteByQuery: removed `_type` from path
+  * Removed core.exists: fully deprecated in ES 7.x
 # v0.7.4 / 2021-06-22
 
 * Updated dependencies